% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle.R
\name{triangle}
\alias{triangle}
\title{Triangle for 2-simplex operations}
\usage{
triangle(labels = c("1", "2", "3"))
}
\arguments{
\item{labels}{String vector of length 3 that names the edges of the
triangle.}
}
\value{
List of size 2 with members \code{coords} and \code{canvas}.
 \code{coords} holds edge coordinates of the 2-simplex, \code{canvas}
 a ggplot2 plot object of the 2-simplex.
}
\description{
Generates a triangle representing the 2-simplex.
}
\examples{
triangle()
}
\author{
Daniel Gebele \email{dngebele@gmail.com}
}
